chore(order-base): changed chrtID data type to int64

+++ order-base/internal/models/item.go
@@ -1,7 +1,7 @@
 package models
 
 type Item struct {
-	ChrtID      int    `json:"chrt_id" db:"chrt_id"`
+	ChrtID      int64    `json:"chrt_id" db:"chrt_id"`
 	TrackNumber string `json:"track_number" db:"track_number"`
 	Price       int    `json:"price" db:"price"`
 	Rid         string `json:"rid" db:"rid"`
